Create ktDataset.py
We propose to include a new dataset on Kamlet-Taft (KT) parameters (input: SMILES; train: KT parameters). These parameters are experimentally measured using Nuclear magnetic resonance (NMR) spectroscopy and the dataset loaded here has been hand curated from several papers. The dataset, we believe, will be broadly useful for the community given the effort in curating it. The proposal to include as a dataset was first made to Bharath. In addition, the dataset .csv has been uploaded to the S3 bucket.